requirements.txt
APScheduler==3.9.1.post1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
filelock==3.9.0
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy
protobuf==3.20.3
PyMySQL==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
schedule==1.1.0
scikit-learn
scipy
six==1.16.0
snscrape
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
Werkzeug==2.2.2
gunicorn